requirements.txt
openai
streamlit
python-dotenv
pydantic
os
